requirements.txt
gradio==5.40.0
huggingface_hub==0.28.1
numpy==1.24.3
pandas==2.0.3
Pillow==11.3.0
scikit_learn==1.3.2
scipy==1.16.1
timm==1.0.19
torch==2.4.1
torchvision==0.19.1
